Change 'Wall' constructor
Given that the 'Maze' class will set up the walls through the use
of copies of a dummy wall it became unnecessary for the 'Wall'
constructor to specify its position. Now it sets its origin to the
middle of the wall instead. The order of the variables for defining
the size of the wall was reversed. This for making the constructed
wall more intuitive by having it extend right when without any
transformations.